chore: Refactor package names in InfrastructureAgent code files
