+++ InfrastructureAgent/model/cpu_metrics.go
@@ -1,0 +1,5 @@
+package oneuptime_infrastructure_agent
+
+type CPUMetrics struct {
+	PercentUsed float64 `json:"percentUsed"`
+}
