+++ InfrastructureAgent/model/disk_metrics.go
@@ -1,0 +1,10 @@
+package oneuptime_infrastructure_agent
+
+type BasicDiskMetrics struct {
+	Total       uint64  `json:"total"`
+	Free        uint64  `json:"free"`
+	Used        uint64  `json:"used"`
+	DiskPath    string  `json:"diskPath"`
+	PercentUsed float64 `json:"percentUsed"`
+	PercentFree float64 `json:"percentFree"`
+}
