+++ InfrastructureAgent/model/infrastructure_metrics.go
@@ -1,0 +1,7 @@
+package oneuptime_infrastructure_agent
+
+type BasicInfrastructureMetrics struct {
+	CpuMetrics    *CPUMetrics         `json:"cpuMetrics"`
+	MemoryMetrics *MemoryMetrics      `json:"memoryMetrics"`
+	DiskMetrics   []*BasicDiskMetrics `json:"diskMetrics"`
+}
