+++ InfrastructureAgent/model/server_process.go
@@ -1,0 +1,7 @@
+package oneuptime_infrastructure_agent
+
+type ServerProcess struct {
+	Pid     int32  `json:"pid"`
+	Name    string `json:"name"`
+	Command string `json:"command"`
+}
